+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-﻿cMeta (0.12.0):
+﻿cMeta (0.13.0):
+  * Added state['deps'] for python packages
+
+cMeta (0.12.0):
   * Added new categories to support CK
 
 cMeta (0.11.0):